python/requirements.txt
# Tushare数据集成依赖
tushare>=1.4.0
pandas>=2.0.0
numpy>=1.24.0

# 数据库依赖
pymysql>=1.1.0
SQLAlchemy>=2.0.0

# 定时任务依赖
APScheduler>=3.10.0

# 配置管理依赖
python-dotenv>=1.0.0

# 日志依赖（Python内置，无需安装）
# logging

# 邮件依赖（Python内置，无需安装）
# smtplib
# email
